// lambda/schoolHandler/src/models/schoolDetails.model.ts
class SchoolDetails {
    SchoolCode?: string;
    SchoolFullName?: string;
    SchoolPreschoolIndicator?: string;
    SchoolLevelSubtypeCode?: string;
    SchoolLevelSubtypeName?: string;
    SchoolPhoneNumber?: string;
    SchoolEmailAddress?: string;
    SchoolSuburbName?: string;
    SchoolPrincipalName?: string;
    SchoolTotalStudents?: number;
    SchoolEmployeeTeachingCount?: number;
}

export { SchoolDetails };